fee api is completed

+++ src/main/java/manager/repository/FeeRepository.java
@@ -4,9 +4,10 @@ import manager.entity.Fee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface FeeRepository extends JpaRepository<Fee, Integer> {
-    List<Fee> findByRoomCategoryId(int customerId);
+public interface FeeRepository extends JpaRepository<Fee, BigInteger> {
+    List<Fee> findByRoomCategoryId(BigInteger customerId);
 }
